+++ lib/src/controllers/pod_video_quality_controller.dart
@@ -20,13 +20,13 @@ class _PodVideoQualityController extends _PodVideoController {
   }) async {
     try {
       podVideoStateChanger(PodVideoState.loading);
-      final _vimeoVideoUrls = await VideoApis.getVimeoVideoQualityUrls(
+      final vimeoVideoUrls = await VideoApis.getVimeoVideoQualityUrls(
         videoId,
         hash,
       );
 
       ///
-      vimeoOrVideoUrls = _vimeoVideoUrls ?? [];
+      vimeoOrVideoUrls = vimeoVideoUrls ?? [];
     } catch (e) {
       rethrow;
     }
@@ -38,11 +38,11 @@ class _PodVideoQualityController extends _PodVideoController {
   ) async {
     try {
       podVideoStateChanger(PodVideoState.loading);
-      final _vimeoVideoUrls =
+      final vimeoVideoUrls =
           await VideoApis.getVimeoPrivateVideoQualityUrls(videoId, httpHeader);
 
       ///
-      vimeoOrVideoUrls = _vimeoVideoUrls ?? [];
+      vimeoOrVideoUrls = vimeoVideoUrls ?? [];
     } catch (e) {
       rethrow;
     }
@@ -51,21 +51,21 @@ class _PodVideoQualityController extends _PodVideoController {
   void sortQualityVideoUrls(
     List<VideoQalityUrls>? urls,
   ) {
-    final _urls = urls;
+    final urls0 = urls;
 
     ///has issues with 240p
-    _urls?.removeWhere((element) => element.quality == 240);
+    urls0?.removeWhere((element) => element.quality == 240);
 
     ///has issues with 144p in web
     if (kIsWeb) {
-      _urls?.removeWhere((element) => element.quality == 144);
+      urls0?.removeWhere((element) => element.quality == 144);
     }
 
     ///sort
-    _urls?.sort((a, b) => a.quality.compareTo(b.quality));
+    urls0?.sort((a, b) => a.quality.compareTo(b.quality));
 
     ///
-    vimeoOrVideoUrls = _urls ?? [];
+    vimeoOrVideoUrls = urls0 ?? [];
   }
 
   ///get vimeo quality `ex: 1080p` url
@@ -127,7 +127,7 @@ class _PodVideoQualityController extends _PodVideoController {
       podVideoStateChanger(PodVideoState.paused);
       podVideoStateChanger(PodVideoState.loading);
       playingVideoUrl = _videoQualityUrl;
-      _videoCtr = VideoPlayerController.network(_videoQualityUrl);
+      _videoCtr = VideoPlayerController.networkUrl(Uri.parse(_videoQualityUrl));
       await _videoCtr?.initialize();
       _videoDuration = _videoCtr?.value.duration ?? Duration.zero;
       _videoCtr?.addListener(videoListner);