+++ lib/src/models/play_video_from.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import '../../pod_player.dart';
 
 class PlayVideoFrom {
@@ -6,7 +8,7 @@ class PlayVideoFrom {
   final PodVideoPlayerType playerType;
   final VideoFormat? formatHint;
   final String? package;
-  final dynamic file;
+  final File? file;
   final List<VideoQalityUrls>? videoQualityUrls;
   final Future<ClosedCaptionFile>? closedCaptionFile;
   final VideoPlayerOptions? videoPlayerOptions;
@@ -14,10 +16,10 @@ class PlayVideoFrom {
   final bool live;
 
   const PlayVideoFrom._({
+    required this.playerType,
     this.live = false,
     this.dataSource,
     this.hash,
-    required this.playerType,
     this.formatHint,
     this.package,
     this.file,
@@ -62,7 +64,7 @@ class PlayVideoFrom {
   ///File Doesnot support web apps
   ///[file] is `File` Datatype import it from `dart:io`
   factory PlayVideoFrom.file(
-    dynamic file, {
+    File file, {
     Future<ClosedCaptionFile>? closedCaptionFile,
     VideoPlayerOptions? videoPlayerOptions,
   }) {