+++ lib/src/widgets/core/overlays/mobile_bottomsheet.dart
@@ -4,26 +4,25 @@ class _MobileBottomSheet extends StatelessWidget {
   final String tag;
 
   const _MobileBottomSheet({
-    Key? key,
     required this.tag,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
     return GetBuilder<PodGetXVideoController>(
       tag: tag,
-      builder: (_podCtr) => Column(
+      builder: (podCtr) => Column(
         mainAxisSize: MainAxisSize.min,
         children: [
-          if (_podCtr.vimeoOrVideoUrls.isNotEmpty)
+          if (podCtr.vimeoOrVideoUrls.isNotEmpty)
             _bottomSheetTiles(
-              title: _podCtr.podPlayerLabels.quality,
+              title: podCtr.podPlayerLabels.quality,
               icon: Icons.video_settings_rounded,
-              subText: '${_podCtr.vimeoPlayingVideoQuality}p',
+              subText: '${podCtr.vimeoPlayingVideoQuality}p',
               onTap: () {
                 Navigator.of(context).pop();
                 Timer(const Duration(milliseconds: 100), () {
-                  showModalBottomSheet(
+                  showModalBottomSheet<void>(
                     context: context,
                     builder: (context) => SafeArea(
                       child: _VideoQualitySelectorMob(
@@ -39,14 +38,18 @@ class _MobileBottomSheet extends StatelessWidget {
               },
             ),
           _bottomSheetTiles(
-            title: _podCtr.podPlayerLabels.loopVideo,
+            title: podCtr.podPlayerLabels.loopVideo,
             icon: Icons.loop_rounded,
             subText: _podCtr.isLooping ? _podCtr.podPlayerLabels.optionEnabled : _podCtr.podPlayerLabels.optionDisabled,
+            subText: podCtr.isLooping
+                ? podCtr.podPlayerLabels.optionEnabled
+                : podCtr.podPlayerLabels.optionDisabled,
             onTap: () {
               Navigator.of(context).pop();
-              _podCtr.toggleLooping();
+              podCtr.toggleLooping();
             },
           ),
+
           if (_podCtr.showPlaybackSpeed)
             _bottomSheetTiles(
               title: _podCtr.podPlayerLabels.playbackSpeed,
@@ -63,7 +66,6 @@ class _MobileBottomSheet extends StatelessWidget {
                         tag: tag,
                         onTap: null,
                       ),
-                    ),
                   );
                 });
               },
@@ -120,25 +122,24 @@ class _VideoQualitySelectorMob extends StatelessWidget {
   final String tag;
 
   const _VideoQualitySelectorMob({
-    Key? key,
     required this.onTap,
     required this.tag,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
-    final _podCtr = Get.find<PodGetXVideoController>(tag: tag);
+    final podCtr = Get.find<PodGetXVideoController>(tag: tag);
     return SingleChildScrollView(
       child: Column(
         mainAxisSize: MainAxisSize.min,
-        children: _podCtr.vimeoOrVideoUrls
+        children: podCtr.vimeoOrVideoUrls
             .map(
               (e) => ListTile(
                 title: Text('${e.quality}p'),
                 onTap: () {
                   onTap != null ? onTap!() : Navigator.of(context).pop();
 
-                  _podCtr.changeVideoQuality(e.quality);
+                  podCtr.changeVideoQuality(e.quality);
                 },
               ),
             )
@@ -153,24 +154,23 @@ class _VideoPlaybackSelectorMob extends StatelessWidget {
   final String tag;
 
   const _VideoPlaybackSelectorMob({
-    Key? key,
     required this.onTap,
     required this.tag,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
-    final _podCtr = Get.find<PodGetXVideoController>(tag: tag);
+    final podCtr = Get.find<PodGetXVideoController>(tag: tag);
     return SingleChildScrollView(
       child: Column(
         mainAxisSize: MainAxisSize.min,
-        children: _podCtr.videoPlaybackSpeeds
+        children: podCtr.videoPlaybackSpeeds
             .map(
               (e) => ListTile(
                 title: Text(e),
                 onTap: () {
                   onTap != null ? onTap!() : Navigator.of(context).pop();
-                  _podCtr.setVideoPlayBack(e);
+                  podCtr.setVideoPlayBack(e);
                 },
               ),
             )
@@ -184,9 +184,8 @@ class _MobileOverlayBottomControlles extends StatelessWidget {
   final String tag;
 
   const _MobileOverlayBottomControlles({
-    Key? key,
     required this.tag,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -196,7 +195,7 @@ class _MobileOverlayBottomControlles extends StatelessWidget {
     return GetBuilder<PodGetXVideoController>(
       tag: tag,
       id: 'full-screen',
-      builder: (_podCtr) => Column(
+      builder: (podCtr) => Column(
         mainAxisSize: MainAxisSize.min,
         children: [
           Row(
@@ -205,11 +204,11 @@ class _MobileOverlayBottomControlles extends StatelessWidget {
               GetBuilder<PodGetXVideoController>(
                 tag: tag,
                 id: 'video-progress',
-                builder: (_podCtr) {
+                builder: (podCtr) {
                   return Row(
                     children: [
                       Text(
-                        _podCtr.calculateVideoDuration(_podCtr.videoPosition),
+                        podCtr.calculateVideoDuration(podCtr.videoPosition),
                         style: const TextStyle(color: itemColor),
                       ),
                       const Text(
@@ -217,7 +216,7 @@ class _MobileOverlayBottomControlles extends StatelessWidget {
                         style: durationTextStyle,
                       ),
                       Text(
-                        _podCtr.calculateVideoDuration(_podCtr.videoDuration),
+                        podCtr.calculateVideoDuration(podCtr.videoDuration),
                         style: durationTextStyle,
                       ),
                     ],
@@ -225,24 +224,28 @@ class _MobileOverlayBottomControlles extends StatelessWidget {
                 },
               ),
               const Spacer(),
-              MaterialIconButton(
-                toolTipMesg: _podCtr.isFullScreen
-                    ? _podCtr.podPlayerLabels.exitFullScreen ?? 'Exit full screen${kIsWeb ? ' (f)' : ''}'
-                    : _podCtr.podPlayerLabels.fullscreen ?? 'Fullscreen${kIsWeb ? ' (f)' : ''}',
+             MaterialIconButton(
+                toolTipMesg: podCtr.isFullScreen
+                    ? podCtr.podPlayerLabels.exitFullScreen ??
+                        'Exit full screen${kIsWeb ? ' (f)' : ''}'
+                    : podCtr.podPlayerLabels.fullscreen ??
+                        'Fullscreen${kIsWeb ? ' (f)' : ''}',
                 color: itemColor,
                 onPressed: () {
-                  if (_podCtr.isOverlayVisible) {
-                    if (_podCtr.isFullScreen) {
-                      _podCtr.disableFullScreen(context, tag);
+                  if (podCtr.isOverlayVisible) {
+                    if (podCtr.isFullScreen) {
+                      podCtr.disableFullScreen(context, tag);
                     } else {
-                      _podCtr.enableFullScreen(tag);
+                      podCtr.enableFullScreen(tag);
                     }
                   } else {
-                    _podCtr.toggleVideoOverlay();
+                    podCtr.toggleVideoOverlay();
                   }
                 },
                 child: Icon(
-                  _podCtr.isFullScreen ? Icons.fullscreen_exit : Icons.fullscreen,
+                  podCtr.isFullScreen
+                      ? Icons.fullscreen_exit
+                      : Icons.fullscreen,
                 ),
               ),
             ],
@@ -250,30 +253,24 @@ class _MobileOverlayBottomControlles extends StatelessWidget {
           GetBuilder<PodGetXVideoController>(
             tag: tag,
             id: 'overlay',
-            builder: (_podCtr) {
-              if (_podCtr.isFullScreen) {
+            builder: (podCtr) {
+              if (podCtr.isFullScreen) {
                 return Padding(
                   padding: const EdgeInsets.fromLTRB(12, 0, 12, 20),
                   child: Visibility(
-                    visible: _podCtr.isOverlayVisible,
+                    visible: podCtr.isOverlayVisible,
                     child: PodProgressBar(
                       tag: tag,
                       alignment: Alignment.topCenter,
-                      podProgressBarConfig: _podCtr.podProgressBarConfig,
+                      podProgressBarConfig: podCtr.podProgressBarConfig,
                     ),
                   ),
                 );
               }
-              return Padding(
-                padding: _podCtr.podProgressBarConfig.padding,
-                child: Padding(
-                  padding: _podCtr.podProgressBarConfig.padding,
-                  child: PodProgressBar(
-                    tag: tag,
-                    alignment: Alignment.bottomCenter,
-                    podProgressBarConfig: _podCtr.podProgressBarConfig,
-                  ),
-                ),
+              return PodProgressBar(
+                tag: tag,
+                alignment: Alignment.bottomCenter,
+                podProgressBarConfig: podCtr.podProgressBarConfig,
               );
             },
           ),