+++ lib/src/widgets/core/overlays/overlays.dart
@@ -4,42 +4,41 @@ class _VideoOverlays extends StatelessWidget {
   final String tag;
 
   const _VideoOverlays({
-    Key? key,
     required this.tag,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
-    final _podCtr = Get.find<PodGetXVideoController>(tag: tag);
-    if (_podCtr.overlayBuilder != null) {
+    final podCtr = Get.find<PodGetXVideoController>(tag: tag);
+    if (podCtr.overlayBuilder != null) {
       return GetBuilder<PodGetXVideoController>(
         id: 'update-all',
         tag: tag,
-        builder: (_podCtr) {
+        builder: (podCtr) {
           ///Custom overlay
-          final _progressBar = PodProgressBar(
+          final progressBar = PodProgressBar(
             tag: tag,
-            podProgressBarConfig: _podCtr.podProgressBarConfig,
+            podProgressBarConfig: podCtr.podProgressBarConfig,
           );
           final overlayOptions = OverLayOptions(
-            podVideoState: _podCtr.podVideoState,
-            videoDuration: _podCtr.videoDuration,
-            videoPosition: _podCtr.videoPosition,
-            showPlaybackSpeed: _podCtr.showPlaybackSpeed,
-            isFullScreen: _podCtr.isFullScreen,
-            isLooping: _podCtr.isLooping,
-            isOverlayVisible: _podCtr.isOverlayVisible,
-            isMute: _podCtr.isMute,
-            autoPlay: _podCtr.autoPlay,
-            currentVideoPlaybackSpeed: _podCtr.currentPaybackSpeed,
-            videoPlayBackSpeeds: _podCtr.videoPlaybackSpeeds,
-            videoPlayerType: _podCtr.videoPlayerType,
-            podProgresssBar: _progressBar,
+            podVideoState: podCtr.podVideoState,
+            videoDuration: podCtr.videoDuration,
+            videoPosition: podCtr.videoPosition,
+            showPlaybackSpeed: podCtr.showPlaybackSpeed,
+            isFullScreen: podCtr.isFullScreen,
+            isLooping: podCtr.isLooping,
+            isOverlayVisible: podCtr.isOverlayVisible,
+            isMute: podCtr.isMute,
+            autoPlay: podCtr.autoPlay,
+            currentVideoPlaybackSpeed: podCtr.currentPaybackSpeed,
+            videoPlayBackSpeeds: podCtr.videoPlaybackSpeeds,
+            videoPlayerType: podCtr.videoPlayerType,
+            podProgresssBar: progressBar,
           );
 
           /// Returns the custom overlay, otherwise returns the default
           /// overlay with gesture detector
-          return _podCtr.overlayBuilder!(overlayOptions);
+          return podCtr.overlayBuilder!(overlayOptions);
         },
       );
     } else {
@@ -47,10 +46,10 @@ class _VideoOverlays extends StatelessWidget {
       return GetBuilder<PodGetXVideoController>(
         tag: tag,
         id: 'overlay',
-        builder: (_podCtr) {
+        builder: (podCtr) {
           return AnimatedOpacity(
             duration: const Duration(milliseconds: 200),
-            opacity: _podCtr.isOverlayVisible ? 1 : 0,
+            opacity: podCtr.isOverlayVisible ? 1 : 0,
             child: Stack(
               fit: StackFit.passthrough,
               children: [