+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.2.1
+- upgraded all dependencies 
+- upgraded `http` lib to `^1.1.0`
+- fix fullscreen issue in IOS
+
+## 0.2.0
+  - upgraded dependencies 
+  - migrate from `wakelock` to `wakelock_plus` PR [#129](https://github.com/newtaDev/pod_player/pull/129)
+  - fixed all lint rules
+  - migrated `VideoPlayerController.network` to `VideoPlayerController.networkUrl`
+  - Breaking:
+    - In `PlayVideoFrom.file` [file] param datatype changed from [dynamic] to [File]
 ## 0.1.5
   - merged PR #103
     - support unlisted vimeo videos